LayoutTests/fast/forms/number/number-validation-message-expected.txt

Test for validationMessage IDL attribute for <input type=number>

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


No message
PASS testIt("", "", "") is ""
Value missing
PASS testIt("", "", "") is "Fill out this field"
Type mismatch
PASS testIt("foo", "", "") is ""
Range overflow
PASS testIt("200", "", "100") is "Value must be less than or equal to 100"
Range underflow
PASS testIt("10", "50", "100") is "Value must be greater than or equal to 50"
Step mismatch
PASS testIt("55", "0", "100", "10") is "Enter a valid value"
PASS successfullyParsed is true

TEST COMPLETE
